Move signal implementations to own files
